LayoutTests/imported/w3c/web-platform-tests/wpt_internal/html/canvas/drawElementImage/privacy/subframe-cross-origin-background-image.https.sub-expected.txt



FAIL subframe loading background images assert_array_equals: Same origin should draw expected property 1 to be 255 but got 0 (expected array [0, 255, 0, 255] got object "0,0,0,0")
